fastapi/requirements.txt
fastapi
uvicorn
Pillow
google-genai
google-adk
python-dotenv
